Add tests for getTotalSkillBonuses
