+++ src/store/Skills/reducers.ts
@@ -1,39 +1,11 @@
 import {
   quickSkillDefinitions,
-  Skill,
   SkillState,
   UPDATE_SKILL_IS_CLASS_SKILL,
   UPDATE_SKILL_MISC_MODIFIER,
   UPDATE_SKILL_RANKS,
 } from './types'
-import { abilityName } from '../Abilities/types'
-import { getAbilityModifiers } from '../Abilities/selectors'
-import { RootActionTypes, RootState } from '../root-reducer'
-import { createSelector } from 'reselect'
-
-export const computeTotalSkillBonus = (skill: Skill, abilityModifier: number): number => {
-  return abilityModifier + skill.ranks + skill.miscModifier + (
-    skill.isClassSkill && skill.ranks >= 1 ? 3 : 0
-  )
-}
-
-export const getSkills = (state: RootState) => state.skills
-
-export type TotalSkillBonuses = {
-  [k in keyof SkillState]: number
-}
-
-export const getTotalSkillBonuses = createSelector([getSkills, getAbilityModifiers], (skills, abilityModifiers) => {
-  const skillNames = Object.keys(skills) as Array<keyof SkillState>
-  return skillNames.reduce((totalSkillsBonuses, currSkillName) => {
-    const baseAbilityName = quickSkillDefinitions[currSkillName].baseAbility as abilityName
-
-    return {
-      ...totalSkillsBonuses,
-      [currSkillName]: computeTotalSkillBonus(skills[currSkillName], abilityModifiers[baseAbilityName]),
-    }
-  }, {} as TotalSkillBonuses)
-})
+import { RootActionTypes } from '../root-reducer'
 
 export const createInitialState = () => {
   const result = {} as SkillState
